+++ docs/changelog.md
@@ -3,6 +3,9 @@
 ## v0.9.1 - Master
 ####* This version is not yet released and is under active development.
 
+###Fixed
+
+  - Correctly reporting architecture
 
 ## v0.9.0 - Master